Refactor job fetching logic: rename jobWord to searchWord for consistency, update related functions and queries to reflect the change
