+++ documentation/jobFetch/4.3.Search Processing & Results Aggregation.md
@@ -33,7 +33,7 @@ sequenceDiagram
 
         Note over JobController: Create concurrent fetch promises<br/>for all search words
         loop For each search word (index i)
-            JobController->>NeonDB: Check if `jobWord` exists in `search_words`
+            JobController->>NeonDB: Check if `searchWord` exists in `search_words`
             activate NeonDB
             alt Word cached in DB
                 NeonDB-->>JobController: Return cached jobs from `jobs` table
@@ -42,9 +42,9 @@ sequenceDiagram
                 JobController->>JobController: Fallback to real search
                 alt i >= 2 AND searchWords.length > 2
                     Note over JobController: Apply throttling:<br/>setTimeout (i-1) * 700ms
-                    JobController->>RealJobSearch: call `rapidAPIfetch(jobWord)`
+                    JobController->>RealJobSearch: call `rapidAPIfetch(searchWord)`
                 else
-                    JobController->>RealJobSearch: call `rapidAPIfetch(jobWord)` (immediate)
+                    JobController->>RealJobSearch: call `rapidAPIfetch(searchWord)` (immediate)
                 end
                 activate RealJobSearch
                 RealJobSearch->>RapidAPI: GET LinkedIn Job Search API
